+++ ports/servoshell/desktop/accelerated_gl_media.rs
@@ -14,41 +14,39 @@ pub(crate) fn setup_gl_accelerated_media(_: RefMut<'_, Device>, _: RefMut<'_, Co
 
 #[cfg(all(target_os = "linux", not(target_env = "ohos")))]
 pub(crate) fn setup_gl_accelerated_media(device: RefMut<'_, Device>, context: RefMut<'_, Context>) {
-    use servo::Servo;
-    use servo::media::{GlContext, NativeDisplay};
+    use servo::{MediaGlContext, MediaNativeDisplay, Servo};
     use surfman::platform::generic::multi::connection::NativeConnection;
     use surfman::platform::generic::multi::context::NativeContext;
 
     let api = api(&device, &context);
     let context = match device.native_context(&context) {
         NativeContext::Default(NativeContext::Default(native_context)) => {
-            GlContext::Egl(native_context.egl_context as usize)
+            MediaGlContext::Egl(native_context.egl_context as usize)
         },
         NativeContext::Default(NativeContext::Alternate(native_context)) => {
-            GlContext::Egl(native_context.egl_context as usize)
+            MediaGlContext::Egl(native_context.egl_context as usize)
         },
-        NativeContext::Alternate(_) => GlContext::Unknown,
+        NativeContext::Alternate(_) => MediaGlContext::Unknown,
     };
     let display = match device.connection().native_connection() {
         surfman::NativeConnection::Default(NativeConnection::Default(connection)) => {
-            NativeDisplay::Egl(connection.0 as usize)
+            MediaNativeDisplay::Egl(connection.0 as usize)
         },
         surfman::NativeConnection::Default(NativeConnection::Alternate(connection)) => {
-            NativeDisplay::X11(connection.x11_display as usize)
+            MediaNativeDisplay::X11(connection.x11_display as usize)
         },
-        surfman::NativeConnection::Alternate(_) => NativeDisplay::Unknown,
+        surfman::NativeConnection::Alternate(_) => MediaNativeDisplay::Unknown,
     };
     Servo::initialize_gl_accelerated_media(display, api, context);
 }
 
 #[cfg(target_os = "windows")]
 pub(crate) fn setup_gl_accelerated_media(device: RefMut<'_, Device>, context: RefMut<'_, Context>) {
-    use servo::Servo;
-    use servo::media::{GlContext, NativeDisplay};
+    use servo::{MediaGlContext, MediaNativeDisplay, Servo};
 
     let api = api(&device, &context);
-    let context = GlContext::Egl(device.native_context(&context).egl_context as usize);
-    let display = NativeDisplay::Egl(device.native_device().egl_display as usize);
+    let context = MediaGlContext::Egl(device.native_context(&context).egl_context as usize);
+    let display = MediaNativeDisplay::Egl(device.native_device().egl_display as usize);
     Servo::initialize_gl_accelerated_media(display, api, context);
 }
 
@@ -56,8 +54,8 @@ pub(crate) fn setup_gl_accelerated_media(device: RefMut<'_, Device>, context: Re
     all(target_os = "linux", not(target_env = "ohos")),
     target_os = "windows"
 ))]
-fn api(device: &RefMut<Device>, context: &RefMut<Context>) -> servo::media::GlApi {
-    use servo::media::GlApi;
+fn api(device: &RefMut<Device>, context: &RefMut<Context>) -> servo::MediaGlApi {
+    use servo::MediaGlApi;
     use surfman::GLApi;
 
     let descriptor = device.context_descriptor(context);
@@ -65,9 +63,9 @@ fn api(device: &RefMut<Device>, context: &RefMut<Context>) -> servo::media::GlAp
     let major = attributes.version.major;
     let minor = attributes.version.minor;
     match device.connection().gl_api() {
-        GLApi::GL if major >= 3 && minor >= 2 => GlApi::OpenGL3,
-        GLApi::GL => GlApi::OpenGL,
-        GLApi::GLES if major > 1 => GlApi::Gles2,
-        GLApi::GLES => GlApi::Gles1,
+        GLApi::GL if major >= 3 && minor >= 2 => MediaGlApi::OpenGL3,
+        GLApi::GL => MediaGlApi::OpenGL,
+        GLApi::GLES if major > 1 => MediaGlApi::Gles2,
+        GLApi::GLES => MediaGlApi::Gles1,
     }
 }
