+++ ports/servoshell/desktop/protocols/resource.rs
@@ -13,13 +13,10 @@ use std::io::BufReader;
 use std::pin::Pin;
 
 use headers::{ContentType, HeaderMapExt};
-use net::fetch::methods::{DoneChannel, FetchContext};
-use net::filemanager_thread::FILE_CHUNK_SIZE;
-use net::protocols::ProtocolHandler;
-use net_traits::ResourceFetchTiming;
-use net_traits::filemanager_thread::RelativePos;
-use net_traits::request::Request;
-use net_traits::response::{Response, ResponseBody};
+use servo::protocol_handler::{
+    DoneChannel, FILE_CHUNK_SIZE, FetchContext, ProtocolHandler, RelativePos, Request,
+    ResourceFetchTiming, Response, ResponseBody,
+};
 use tokio::sync::mpsc::unbounded_channel;
 
 #[derive(Default)]
