+++ ports/servoshell/parser.rs
@@ -5,8 +5,7 @@
 #[cfg(not(any(target_os = "android", target_env = "ohos")))]
 use std::path::{Path, PathBuf};
 
-use servo::net_traits::pub_domains::is_reg_domain;
-use servo::servo_url::ServoUrl;
+use servo::{ServoUrl, is_reg_domain};
 
 #[cfg(not(any(target_os = "android", target_env = "ohos")))]
 pub fn parse_url_or_filename(cwd: &Path, input: &str) -> Result<ServoUrl, ()> {
